+++ src/alley/wp/alleyvate/load.php
@@ -34,6 +34,10 @@ function load(): void {
 			new Features\Clean_Admin_Bar(),
 		),
 		new Feature(
+			'disable_apple_news_non_prod_push',
+			new Features\Disable_Apple_News_Non_Prod_Push(),
+		),
+		new Feature(
 			'disable_attachment_routing',
 			new Features\Disable_Attachment_Routing(),
 		),